remove context and extract Secret to SecretActions
